backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-filter==24.2
kombu==5.5.3
mysqlclient==2.2.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
python-decouple==3.8
redis==5.0.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
gunicorn==20.1.0
django-celery-results
